UBSan: Fix byte overflow in objIndexData::max_struct
Change from byte to short to accommodate values 0-250+ and -1 sentinel.

The byte type (range -128 to 127) was insufficient for max_struct values
which can exceed 127 (documented as supporting 0-250+). When values
greater than 127 were loaded from the database (stored as int(11)), they
would overflow and wrap to negative values, causing:

- Incorrect object durability calculations
- Potential sorting errors in immortal commands
- Shop pricing errors when comparing cur_struct to max_struct

The related TObj field max_struct_points was already a short, making
this inconsistency particularly error-prone. This change brings the
template field in line with the instance field.

+++ code/code/sys/db.h
@@ -280,7 +280,7 @@ class objIndexData : public indexData {
   public:
     extraDescription* ex_description;  // extra descriptions
     objAffData affected[MAX_OBJ_AFFECT];
-    byte max_struct;
+    short max_struct;  // Can be -1 sentinel or positive (0-250+)
     short armor;
     unsigned int where_worn;
     ubyte itemtype;
